modified building exporter to directly export boundary surfaces
